refactor: TableGroup 일급 컬렉션 수정
